Add liquidate logic to status cell
