+++ src/pages/OffersPage/components/ActiveOffersTable/hooks/useLenderLoansAndOffers.ts
@@ -6,6 +6,7 @@ import { produce } from 'immer'
 import { create } from 'zustand'
 
 import { Offer, fetchLenderLoansAndOffers } from '@banx/api/core'
+import { useOffersOptimistic } from '@banx/store'
 
 interface HiddenNftsMintsState {
   mints: string[]
@@ -23,54 +24,11 @@ const useHiddenNftsMints = create<HiddenNftsMintsState>((set) => ({
   },
 }))
 
-interface OptimisticOffersState {
-  offers: Offer[]
-  addOffer: (offer: Offer) => void
-  findOffer: (offerPubkey: string) => Offer | null
-  removeOffer: (offer: Offer) => void
-  updateOffer: (offer: Offer) => void
-}
-
-const useOptimisticOffers = create<OptimisticOffersState>((set, get) => ({
-  offers: [],
-  addOffer: (offer) => {
-    set(
-      produce((state: OptimisticOffersState) => {
-        state.offers.push(offer)
-      }),
-    )
-  },
-  findOffer: (offerPubkey) => {
-    const { offers } = get()
-    return offers.find(({ publicKey }) => publicKey === offerPubkey) ?? null
-  },
-  removeOffer: (offer) => {
-    set(
-      produce((state: OptimisticOffersState) => {
-        state.offers = state.offers.filter(({ publicKey }) => publicKey !== offer.publicKey)
-      }),
-    )
-  },
-  updateOffer: (offer: Offer) => {
-    const { findOffer } = get()
-    const offerExists = !!findOffer(offer.publicKey)
-
-    offerExists &&
-      set(
-        produce((state: OptimisticOffersState) => {
-          state.offers = state.offers.map((existingOffer) =>
-            existingOffer.publicKey === offer.publicKey ? offer : existingOffer,
-          )
-        }),
-      )
-  },
-}))
-
 export const useLenderLoansAndOffers = () => {
   const { publicKey } = useWallet()
   const publicKeyString = publicKey?.toBase58() || ''
 
-  const { offers: optimisticOffers, findOffer, updateOffer, addOffer } = useOptimisticOffers()
+  const { offers: optimisticOffers, findOffer, updateOffer, addOffer } = useOffersOptimistic()
   const { mints, addMints } = useHiddenNftsMints()
 
   const { data, isLoading } = useQuery(
@@ -88,6 +46,7 @@ export const useLenderLoansAndOffers = () => {
     if (!data?.nfts) {
       return []
     }
+
     return data.nfts.filter(({ nft }) => !mints.includes(nft.mint))
   }, [data, mints])
